// graph/client/src/app/styles-graph/palette.ts
export enum NrwlPalette {
  blue_500 = 'hsla(217, 91%, 60%, 1)',
  blue_600 = 'hsla(221, 83%, 53%, 1)',
  sky_500 = 'hsla(199, 89%, 48%, 1)',
  sky_600 = 'hsla(200, 98%, 39%, 1)',
  pink_400 = 'hsla(329, 86%, 70%, 1)',
  pink_500 = 'hsla(330, 81%, 60%, 1)',
  fuchsia_500 = 'hsla(292, 84%, 61%, 1)',
  fuchsia_600 = 'hsla(293, 69%, 49%, 1)',
  fuchsia_700 = 'hsla(295, 72%, 40%, 1)',
  fuchsia_800 = 'hsla(295, 70%, 33%, 1)',
  slate_50 = 'hsla(210, 40%, 98%, 1)',
  slate_100 = 'hsla(210, 40%, 96%, 1)',
  slate_200 = 'hsla(214, 32%, 91%, 1)',
  slate_300 = 'hsla(213, 27%, 84%, 1)',
  slate_400 = 'hsla(213, 27%, 84%, 1)',
  slate_500 = 'hsla(215, 16%, 47%, 1)',
  slate_600 = 'hsla(215, 19%, 35%, 1)',
  slate_700 = 'hsla(215, 25%, 27%, 1)',
  slate_800 = 'hsla(217, 33%, 17%, 1)',
  white = '#fff',
}